requirements.txt
asgiref==3.5.2
autopep8==1.6.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==37.0.4
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==4.0.5
django-allauth==0.51.0
django-braces==1.15.0
django-cors-headers==3.13.0
django-filter==22.1
django-model-utils==4.2.0
django-phonenumber-field==6.3.0
django-phonenumbers==1.0.1
django-star-ratings==0.9.2
django-stdimage==6.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.3
oauthlib==3.2.0
phonenumbers==8.12.50
Pillow==9.1.1
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
pyotp==2.6.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
sqlparse==0.4.2
swapper==1.3.0
toml==0.10.2
tzdata==2022.1
urllib3==1.26.9
django-multiselectfield==0.1.12